--- test/fixtures/plugin-sdk-package-consumer/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "plugin-sdk-package-consumer",
-  "private": true,
-  "type": "module"
-}
