--- test/fixtures/plugin-sdk-package-consumer/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["DOM", "DOM.Iterable", "ES2023"],
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "noEmit": true,
-    "skipLibCheck": false,
-    "strict": true,
-    "types": []
-  },
-  "include": ["index.ts"]
-}
